Create LICENSE
Identify the goal of open sourcing the software. 

Identify the (potential) impact of releasing the software under an open source license: 

For Logiqs 

For our partners and customers 

For developers 

Suggest and motivate a suitable license. 

Describe if and how we will accept contributions from outside Logiqs. 

Share your proposal with Marijn 

Receive an OK. 

Prepare a pull request with the chosen license. 

Open this pull request and persuade either Gert-Jan or Reyer to merge the pull request. This will serve as formal prove that Logiqs as a legal entity backs open sourcing this piece of software.